# Backend/src/Aurigma.DirectMail.Sample.TypeScriptApiClient/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Direct Mail API",
    "version": "v1"
  },
  "paths": {
...
    "/api/direct-mail/v1/previews/design-info": {
      "get": {
        "tags": [
          "Preview"
        ],
        "summary": "Returns a design info.",
        "operationId": "Preview_GetDesignInfo",
        "parameters": [
          {
            "name": "LineItemId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DesignInfoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesignInfoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesignInfoDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/direct-mail/v1/previews/proof": {
      "post": {
        "tags": [
          "Preview"
        ],
        "summary": "Returns a proof file.",
        "operationId": "Preview_RenderDesignProof",
        "requestBody": {
          "description": "Request body.",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ProofRequestModel"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProofRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProofRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProofRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/direct-mail/v1/previews/preview": {
      "post": {
        "tags": [
          "Preview"
        ],
        "summary": "Returns a preview file.",
        "operationId": "Preview_RenderDesignPreview",
        "requestBody": {
          "description": "Request body.",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequestModel"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/direct-mail/v1/previews/proofs-zip": {
      "post": {
        "tags": [
          "Preview"
        ],
        "summary": "Returns an archive with proofs.",
        "operationId": "Preview_DownloadProofsArchive",
        "requestBody": {
          "description": "Request body.",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ProofsZipRequestModel"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProofsZipRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProofsZipRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProofsZipRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConflictDto": {
        "type": "object",
        "properties": {
          "propertyName": {
            "type": "string",
            "nullable": true
          },
          "propertyValue": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DesignInfoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Design ID.",
            "nullable": true
          },
          "surfaceCount": {
            "type": "integer",
            "description": "Design surfaces count.",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "PreviewConfigRequestModel": {
        "type": "object",
        "properties": {
          "width": {
            "type": "integer",
            "description": "Preview width.",
            "format": "int32"
          },
          "height": {
            "type": "integer",
            "description": "Preview height.",
            "format": "int32"
          },
          "surfaceIndex": {
            "type": "integer",
            "description": "Design's surface index.",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PreviewRequestModel": {
        "type": "object",
        "properties": {
          "recipientId": {
            "type": "string",
            "description": "Recipient id.",
            "format": "uuid",
            "nullable": true
          },
          "lineItemId": {
            "type": "string",
            "description": "Current line item id.",
            "format": "uuid"
          },
          "config": {
            "$ref": "#/components/schemas/PreviewConfigRequestModel"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ProofRequestConfigModel": {
        "type": "object",
        "properties": {
          "width": {
            "type": "integer",
            "description": "Preview width.",
            "format": "int32"
          },
          "height": {
            "type": "integer",
            "description": "Preview height.",
            "format": "int32"
          },
          "surfaceIndex": {
            "type": "integer",
            "description": "Design's surface index.",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Rendering preview config."
      },
      "ProofRequestModel": {
        "type": "object",
        "properties": {
          "recipientId": {
            "type": "string",
            "description": "Recipient id.",
            "format": "uuid"
          },
          "lineItemId": {
            "type": "string",
            "description": "Current line item id.",
            "format": "uuid"
          },
          "config": {
            "$ref": "#/components/schemas/ProofRequestConfigModel"
          }
        },
        "additionalProperties": false
      },
      "ProofsZipRequestConfigModel": {
        "type": "object",
        "properties": {
          "width": {
            "type": "integer",
            "description": "Preview width.",
            "format": "int32"
          },
          "height": {
            "type": "integer",
            "description": "Preview height.",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Rendeign preview config."
      },
      "ProofsZipRequestModel": {
        "type": "object",
        "properties": {
          "lineItemId": {
            "type": "string",
            "description": "Current line item id.",
            "format": "uuid"
          },
          "config": {
            "$ref": "#/components/schemas/ProofsZipRequestConfigModel"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}